refactor: change the class of the components -search-view-, card-list to functional and create a custom hook -use-loading-
